[UsdImagingGL] Removed most direct use of GL from UsdImagingGL
The remaining use of GL is by UsdImagingGLLegacyEngine and in the
implementation of UsdImagingGLEngine::SetCameraStateFromOpenGL()

The deleted GL state settings have been superseded by render pass
state in HdSt and HgiGL.

Also, we no longer explicitly set the current draw framebuffer
binding as the presentation output in the task controller since
the fallback behavior in HgiInteropOpenGL is to present to the
current draw framebuffer.

(Internal change: 2239200)
